# ebpf: CPU and memory optimization in `findKernelFuncs`